# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is berry im server.",
        "title": "Berry IM API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9191",
    "basePath": "/api/v1",
    "paths": {
        "/ucenter/create": {
            "post": {
                "description": "创建用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ucenter"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "webReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ucenter.WebUserCreatedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ucenter.UserCreatedResultVO"
                        }
                    }
                }
            }
        },
        "/ucenter/getUserById/{uid}": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ucenter"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ucenter.UserBasicInfoResultVO"
                        }
                    }
                }
            }
        },
        "/ucenter/update/{uid}": {
            "put": {
                "description": "更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ucenter"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "webReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ucenter.WebUserUpdatedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ucenter.UserUpdatedResultVO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ucenter.UserBasicInfoResultVO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "ucenter.UserCreatedResultVO": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "string"
                }
            }
        },
        "ucenter.UserUpdatedResultVO": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "string"
                }
            }
        },
        "ucenter.WebUserCreatedRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "ucenter.WebUserUpdatedRequest": {
            "type": "object",
            "required": [
                "clientIp",
                "clientPort",
                "deviceInfo",
                "email",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "clientIp": {
                    "type": "string"
                },
                "clientPort": {
                    "type": "string"
                },
                "deviceInfo": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}